data: DecimalField for storage, confidence constraints, file_size fix
- Change storage_used_mb from FloatField to DecimalField(12,4) to
  eliminate floating-point precision drift
- Add MinValueValidator(0.0) and MaxValueValidator(1.0) to confidence_score
- Fix file_size_bytes to always update when audio_file has .size attribute
  (was only setting on initial upload)
- Update storage arithmetic to use Decimal throughout views and serializers
- 45 tests pass

+++ backend/apps/users/models.py
@@ -34,7 +34,7 @@ class UserProfile(models.Model):
         default='high'
     )
     storage_quota_mb = models.PositiveIntegerField(default=1000)
-    storage_used_mb = models.FloatField(default=0.0)
+    storage_used_mb = models.DecimalField(max_digits=12, decimal_places=4, default=0)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
